Yeah I fucked up something...

+++ app/models/drop.rb
@@ -33,11 +33,14 @@ class Drop
     match_id  = response.data.match_id || response.strategy.match_data.to_s
     
     drop = Drop.find_or_initialize_by(match_id: match_id)
-    if drop.visitable?
-      drop.set_data(response)
-      drop
-    else
-      drop.destroy
+    if drop
+      if drop.visitable?
+        drop.set_data(response)
+        drop
+      else
+        drop.destroy
+        nil
+      end
       nil
     end
   end
